+++ tests/testthat/test-unite.R
@@ -71,3 +71,31 @@ test_that("validates its inputs", {
     unite(df, "z", x:y, na.rm = 1)
   })
 })
+
+test_that("returns an empty string column for empty selections (#1548)", {
+  # i.e. it returns the initial value that would be used in a reduction algorithm
+
+  x <- tibble(
+    x = c("x", "y", "z"),
+    y = c(1, 2, 3)
+  )
+
+  out <- unite(x, "new", all_of(c()))
+
+  expect_identical(names(out), c("x", "y", "new"))
+  expect_identical(out$new, c("", "", ""))
+})
+
+test_that("works with 0 column data frames and empty selections (#1570)", {
+  x <- tibble(.rows = 2L)
+
+  # No `...` implies "unite all the columns"
+  out <- unite(x, "new")
+  expect_identical(names(out), "new")
+  expect_identical(out$new, c("", ""))
+
+  # Empty selection
+  out <- unite(x, "new", all_of(names(x)))
+  expect_identical(names(out), "new")
+  expect_identical(out$new, c("", ""))
+})